ConsoleFormat is value type.
